+++ src/Server/Game/GameServer.cs
@@ -14,6 +14,7 @@ using Imlight.Common.Structures;
 using Imlight.Common.Utilities;
 using Imlight.Common.Cryptography;
 using Imlight.Server.Database;
+using Imlight.Server.Login.Models;
 using Imlight.Server.Shared.Networking;
 using Imlight.Server.Shared.Packets;
 using WizUnraveler.IO;
