+++ src/Server/Imlight.Server.csproj
@@ -16,4 +16,8 @@
         <ProjectReference Include="..\Common\Imlight.Common.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Game\Models\" />
+    </ItemGroup>
+
 </Project>
