+++ src/Server/Login/Models/AuthLevel.cs
@@ -3,13 +3,7 @@
  * Proprietary and confidential.
  */
 
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Imlight.Server.Database
+namespace Imlight.Server.Login.Models
 {
     public enum AuthLevel
     {
