+++ src/Server/Shared/Packets/ACCOUNT_104_PROTOCOL.cs
@@ -5,6 +5,7 @@
 
 using Imlight.Server.Shared.Networking;
 using Imlight.Server.Database;
+using Imlight.Server.Login.Models;
 
 namespace Imlight.Server.Shared.Packets
 {
