+++ src/Server/Shared/Services/AccountService.cs
@@ -11,6 +11,7 @@ using System.Text;
 using System.Threading.Tasks;
 using WizUnraveler.Cache;
 using Imlight.Server.Database;
+using Imlight.Server.Login.Models;
 using Imlight.Server.Shared.Networking;
 using Imlight.Server.Shared.Packets;
